requirements.txt
Django==2.2.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-phone-field==1.8.1
fontawesomefree==6.0.0
phonenumbers==8.12.46
phonenumberslite==8.12.46
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
